Updated schema

+++ src/musicplayer/schema.sql
@@ -1,13 +1,18 @@
-CREATE TABLE music(id int primary key,
+
+CREATE TABLE music(
+id number(6) primary key,
 title varchar(100),
 album varchar(50),
 artist varchar(50),
 genre varchar(50),
 duration varchar(10));
+
 --increase size of name next time
 
-CREATE TABLE files(id int primary key,
+CREATE TABLE files(
+id number(6) primary key,
 file varchar(50));
 
-CREATE TABLE queue(id int,
+CREATE TABLE queue(
+id number(6),
 name varchar(500));
